Move schedule test to a unit test
While this test might not be extremely useful this change removes
the dead code from init.c and schedule.c and moves it to a unit
test.

Acked-by: Frank Lichtenheld <frank@lichtenheld.com>
Gerrit URL: https://gerrit.openvpn.net/c/openvpn/+/1828
Message-Id: <20260730101616.28738-1-frank@lichtenheld.com>
URL: https://www.mail-archive.com/openvpn-devel@lists.sourceforge.net/msg38056.html
