+++ src/openvpn/otime.h
@@ -150,80 +150,6 @@ tv_add(struct timeval *dest, const struct timeval *src)
     }
 }
 
-static inline bool
-tv_lt(const struct timeval *t1, const struct timeval *t2)
-{
-    if (t1->tv_sec < t2->tv_sec)
-    {
-        return true;
-    }
-    else if (t1->tv_sec > t2->tv_sec)
-    {
-        return false;
-    }
-    else
-    {
-        return t1->tv_usec < t2->tv_usec;
-    }
-}
-
-static inline bool
-tv_le(const struct timeval *t1, const struct timeval *t2)
-{
-    if (t1->tv_sec < t2->tv_sec)
-    {
-        return true;
-    }
-    else if (t1->tv_sec > t2->tv_sec)
-    {
-        return false;
-    }
-    else
-    {
-        return t1->tv_usec <= t2->tv_usec;
-    }
-}
-
-static inline bool
-tv_ge(const struct timeval *t1, const struct timeval *t2)
-{
-    if (t1->tv_sec > t2->tv_sec)
-    {
-        return true;
-    }
-    else if (t1->tv_sec < t2->tv_sec)
-    {
-        return false;
-    }
-    else
-    {
-        return t1->tv_usec >= t2->tv_usec;
-    }
-}
-
-static inline bool
-tv_gt(const struct timeval *t1, const struct timeval *t2)
-{
-    if (t1->tv_sec > t2->tv_sec)
-    {
-        return true;
-    }
-    else if (t1->tv_sec < t2->tv_sec)
-    {
-        return false;
-    }
-    else
-    {
-        return t1->tv_usec > t2->tv_usec;
-    }
-}
-
-static inline bool
-tv_eq(const struct timeval *t1, const struct timeval *t2)
-{
-    return t1->tv_sec == t2->tv_sec && t1->tv_usec == t2->tv_usec;
-}
-
 static inline void
 tv_delta(struct timeval *dest, const struct timeval *t1, const struct timeval *t2)
 {
